Refactor Exceptions Refactored exception handling code and tests
