--- src/main/java/seedu/duke/command/repeatexception/InvalidEventListTypeException.java
@@ -1,8 +1,0 @@
-package seedu.duke.command.repeatexception;
-
-public class InvalidEventListTypeException extends InvalidTypeException {
-
-    public InvalidEventListTypeException(String wrongEvent) {
-        super("event type", wrongEvent, "personal, timetable, zoom");
-    }
-}
