--- src/main/java/seedu/duke/command/repeatexception/InvalidTimeUnitException.java
@@ -1,8 +1,0 @@
-package seedu.duke.command.repeatexception;
-
-public class InvalidTimeUnitException extends InvalidTypeException {
-
-    public InvalidTimeUnitException(String wrongTimeUnit) {
-        super("time unit", wrongTimeUnit, "daily, weekly, monthly");
-    }
-}
