--- src/main/java/seedu/duke/command/repeatexception/InvalidTypeException.java
@@ -1,22 +1,0 @@
-package seedu.duke.command.repeatexception;
-
-public class InvalidTypeException extends Exception {
-
-    private String typeCategory;
-    private String typeItem;
-    private String validType;
-    private String message;
-
-    public InvalidTypeException(String typeCategory, String typeItem, String validType) {
-        this.typeCategory = typeCategory;
-        this.typeItem = typeItem;
-        this.validType = validType;
-
-        message = typeItem + " is not a valid " + typeCategory + ". " + "Valid types are: " + validType;
-    }
-
-    @Override
-    public String getMessage() {
-        return message;
-    }
-}
