--- src/main/java/seedu/duke/command/repeatexception/WrongNumberOfArgumentsException.java
@@ -1,16 +1,0 @@
-package seedu.duke.command.repeatexception;
-
-public class WrongNumberOfArgumentsException extends Exception {
-
-    private String message;
-
-    public WrongNumberOfArgumentsException(String errorMessage) {
-        message = errorMessage;
-    }
-
-    @Override
-    public String getMessage() {
-        return this.message;
-    }
-
-}
