+++ src/main/java/seedu/duke/event/Repeat.java
@@ -1,7 +1,6 @@
 package seedu.duke.event;
 
-import seedu.duke.command.repeatexception.InvalidTimeUnitException;
-import seedu.duke.command.repeatexception.InvalidTypeException;
+import seedu.duke.exception.InvalidTimeUnitException;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -23,7 +22,8 @@ public class Repeat {
      * @param repeatType String indicating how far to advance each date for each repetition
      * @param count Integer containing the number of repetition to makes
      */
-    public Repeat(LocalDate startDate, LocalTime startTime, String repeatType, int count) throws InvalidTypeException {
+    public Repeat(LocalDate startDate, LocalTime startTime, String repeatType, int count)
+            throws InvalidTimeUnitException {
         for (int i = 1; i <= count; i++) {
             LocalDate repeatDate;
             if (repeatType.equals(MONTHLY)) {
