+++ src/main/java/seedu/duke/exception/WrongNumberOfArgumentsException.java
@@ -1,0 +1,13 @@
+package seedu.duke.exception;
+
+public class WrongNumberOfArgumentsException extends DukeException {
+
+    private String message;
+
+    public WrongNumberOfArgumentsException(String errorMessage) {
+
+        super(errorMessage);
+    }
+
+
+}
